Check has been completely refactored, I fixed titles for the test classes of check and checkmate as well as fixed a few styling issues
