+++ src/UseCases/Checkmate/Check.java
@@ -1,47 +1,205 @@
 package UseCases.Checkmate;
 
 import Entities.*;
+import UseCases.*;
 
 import java.util.ArrayList;
+import java.util.Arrays;
+
+
 
 /**
  * This class is responsible for determining whether a given King chess piece is in Check or not.
  * Recall that a King is in check when it is in a viable position to be "captured" by an opposing
  * player's piece.
  */
-public class Check {
+public class Check{
 
     // Instance attribute that specifies the positions on the chess board that lead up to the
     // opposing chess piece that is putting the King in Check.
     private int[][] positions;
 
-    /**
-     * This method takes in a king chess piece as well as the GameState it belongs to and determines whether
-     * that King chess piece is in check or not. Recall that a King is in check when it is in a viable position
-     * to be "captured" by an opposing player's piece.
-     *
-     * @return - true, if the King is in check
-     *         - false, otherwise
-     */
+    // Instance attribute that contains the opposing piece the King is being put in
+    // check by
+    private ChessPiece checkedBy;
+
     public boolean isKingInCheck(King king, ChessPiece[][] board){
 
+        // Error checking
         if (king == null){
             return false;
         }
 
-        return checkLeftRight(king, board) || checkForwardsBackwards(king, board) ||
-                checkDiagonals(king, board) || checkKnight(king, board) || checkPawn(king, board);
+        // Loop through each position in our chessboard, if there is an opposing chess
+        // piece in any position, check to see if it has our given King in check.
+        for(int i = 0; i < 8; i++){
+            for(int y = 0; y < 8; y++) {
+                if (board[i][y] != null && (!(board[i][y].getColor().equals(king.getColor())) &&
+                        (!(board[i][y].getLetter() == Character.toChars(0x0198)[0]) &&
+                                !(board[i][y].getLetter() == Character.toChars(0x0199)[0])))){
+
+                    // Generate list of all valid moves, if any of those moves correspond with the King's
+                    // Location, we know that the king is in check
+                    ChessPiece piece = board[i][y];
+                    if (hasKingInCheck(king, piece, board)) {
+                        return true;
+                    }
+                }
+            }
+        }
+        return false;
+    }
+
+    /**
+     * A helper method that returns True if there is an opposing piece that can viably attack the
+     * given King chess piece, and False otherwise.
+     */
+    private boolean hasKingInCheck(King king, ChessPiece piece, ChessPiece[][] board){
+        int[][] validMoves = generateValidMoves(piece, board);
+
+        // If the king's position corresponds to a position in valid moves,
+        // we know that the King is in check by this piece.
+        int[] pos = {king.getRow(), king.getColumn()};
+        for(int[] opposingPos : validMoves){
+            if(Arrays.equals(opposingPos, pos)){
+                this.checkedBy = piece;
+                checkedPositions(king);
+                return true;
+            }
+        }
+        return false;
+    }
+
+    /**
+     * A helper method that takes in a chess piece and a chess board and returns a
+     * list of valid moves corresponding to that piece.
+     */
+    private int[][] generateValidMoves(ChessPiece piece, ChessPiece[][] board){
+        CheckerGenerator checkerGen = new CheckerGenerator();
+        CheckPlayerMove check = checkerGen.generateChecker(piece);
+        return getValidMoves(check, piece, board);
+    }
+
+    /**
+     * This method is only called when the given King is in check. It creates a 2D
+     * integer array that specifies the positions on the chess board that lead up to the
+     * opposing chess piece that is putting the King in Check. It then sets that 2D array
+     * to our instance attribute, positions.
+     */
+    private void checkedPositions(King king){
+
+       // If checkedBy is a pawn or knight, the only way to escape check is by capturing that
+       // opposing chess piece.
+       if(this.checkedBy.getLetter() == 'p' || this.checkedBy.getLetter() == 'P' ||
+               this.checkedBy.getLetter() == 'k' || this.checkedBy.getLetter() == 'K'){
+           this.positions = new int[][] {{this.checkedBy.getRow(), this.checkedBy.getColumn()}};
+           return;
+       }
+       // Otherwise, if checkedBy is a Queen, Rook or Bishop, the only way to escape check is by
+       // through either capturing checkedBy, or by blocking its path of attack.
+       ArrayList<int[]> arr = new ArrayList<>();
+       int[] checkedByPos = new int[] {this.checkedBy.getRow(), this.checkedBy.getColumn()};
+       int[] pos = {this.checkedBy.getRow() - king.getRow(), this.checkedBy.getColumn() - king.getColumn()};
+       if(diagonalPositions(arr, king.getRow(), king.getColumn(), checkedByPos, pos)){
+           return;
+       }
+       straightPositions(arr, king.getRow(), king.getColumn(), checkedByPos, pos);
+    }
+
+    /**
+     * If checkedBy is a chess piece than can move diagonally, then this method is responsible
+     * for returning true when all the positions of it's capturing path have been stored in instance
+     * attribute, positions. Returns false otherwise.
+     */
+    private boolean diagonalPositions(ArrayList<int[]> arr, int row, int col, int[] checkedByPos, int[] pos){
+        // Determines the direction the opposing Queen or Bishop is in
+        // and determines this.positions based off of it.
+        if(pos[0] < 0 && pos[1] < 0){
+            return positionExp(row, col, checkedByPos, arr, -1, -1);
+        }
+        else if(pos[0] < 0 && pos[1] > 0){
+            return positionExp(row, col, checkedByPos, arr, -1, 1);
+        }
+        else if(pos[0] > 0 && pos[1] < 0){
+            return positionExp(row, col, checkedByPos, arr, 1, -1);
+        }
+        else if(pos[0] > 0 && pos[1] > 0){
+            return positionExp(row, col, checkedByPos, arr, 1, 1);
+        }
+        else{
+            return false;
+        }
+    }
+
+    /**
+     * If checkedBy is a chess piece than can move in straights, then this method is responsible
+     * for returning true when all the positions of it's capturing path have been stored in instance
+     * attribute, positions. Returns false otherwise.
+     */
+    private void straightPositions(ArrayList<int[]> arr, int row, int col, int[] checkedByPos, int[] pos){
+        // Determines the direction the opposing Queen or Rook is in
+        // and determines this.positions based off of it.
+        if(pos[0] < 0 && pos[1] == 0){
+            positionExp(row, col, checkedByPos, arr, -1, 0);
+        }
+        else if(pos[0] == 0 && pos[1] > 0){
+            positionExp(row, col, checkedByPos, arr, 0,  1);
+        }
+        else if(pos[0] == 0 && pos[1] < 0){
+            positionExp(row, col, checkedByPos, arr, 0, -1);
+        }
+        else if(pos[0] > 0 && pos[1] == 0){
+            positionExp(row, col, checkedByPos, arr, 1, 0);
+        }
+    }
+
+    /**
+     * This is a helper function used to append positions within the path that the King is
+     * put in check from, to an Arraylist, arr.
+     * @param inc1 - increment used for the row attribute within our while loop
+     * @param inc2 - increment used for the col attribute within our while loop
+     */
+    private boolean positionExp(int row, int col, int[] pos, ArrayList<int[]> arr, int inc1, int inc2){
+        do {
+
+            row = row + inc1;
+            col = col + inc2;
+
+            arr.add(new int[] {row, col});
+
+        } while (row != pos[0] && col != pos[1]);
+        this.positions = toArrayMoves(arr);
+        return true;
+    }
+
+    /**
+     * Helper method used to generate the 2D list of valid moves for each chess piece.
+     */
+    private int[][] getValidMoves(CheckPlayerMove moves, ChessPiece piece, ChessPiece[][] board){
+        if(moves instanceof CheckPawnMove){
+            return ((CheckPawnMove) moves).validMoves(piece, board);
+        }
+        if(moves instanceof CheckKnightMove){
+            return ((CheckKnightMove) moves).validMoves(piece, board);
+        }
+        if(moves instanceof CheckQueenMove){
+            return ((CheckQueenMove) moves).validMoves(piece, board);
+        }
+        if(moves instanceof CheckBishopMove){
+            return ((CheckBishopMove) moves).validMoves(piece, board);
+        }
+        return ((CheckRookMove) moves).validMoves(piece, board);
     }
 
     /**
      * This is a helper method that takes in an array list of integer arrays and
      * transforms its items into a 2d array of the same size, and then returns that array.
      */
-    private int[][] toArrayMoves(ArrayList arr){
+    private int[][] toArrayMoves(ArrayList<int[]> arr){
         int[][] newArr = new int[arr.size()][];
 
         for(int i = 0; i < arr.size(); i ++){
-            newArr[i] = (int[]) arr.get(i);
+            newArr[i] = arr.get(i);
         }
         return newArr;
     }
@@ -52,369 +210,5 @@ public class Check {
      */
     public int[][] getPositions(){
         return this.positions;
-    }
-
-    /**
-     * This method is used to determine whether there is an opposing Queen or Rook that
-     * can capture the King piece from the Left or Right.
-     *
-     * @return - true, if there is an opposing Queen or Rook that can capture the King
-     *         - false, otherwise
-     */
-    private boolean checkLeftRight(King king, ChessPiece[][] board){
-        int col = king.getColumn();
-        int row = king.getRow();
-        ArrayList<int[]> pos = new ArrayList<>();
-
-        // Checking for any Rooks or Queens to the Right of the king
-        while(col + 1 != 8){
-            col++;
-            pos.add(new int[]{row, col});
-            int x = hasOpposingQueenRook(board, row, col, king.getColor());
-            if (x == 1) {
-                this.positions = toArrayMoves(pos);
-                return true;
-            } else if (x == -1) {
-                break;
-            }
-        }
-
-        pos = new ArrayList<>();
-        col = king.getColumn();
-
-        //Checking for any Rooks or Queens to the Left of the King
-        while(col - 1 != -1){
-            col--;
-            pos.add(new int[]{row, col});
-            int x = hasOpposingQueenRook(board, row, col, king.getColor());
-            if (x == 1) {
-                this.positions = toArrayMoves(pos);
-                return true;
-            } else if (x == -1) {
-                break;
-            }
-        }
-        return false;
-    }
-
-    /**
-     * This method is used to determine whether there is an opposing Queen or Rook that
-     * can capture the King piece from directly above or behind it.
-     *
-     * @return - true, if there is an opposing Queen or Rook that can capture the King
-     *         - false, otherwise
-     */
-    private boolean checkForwardsBackwards(King king, ChessPiece[][] board){
-        int col = king.getColumn();
-        int row = king.getRow();
-        ArrayList<int[]> pos = new ArrayList<>();
-
-        // Checking for any Queens or Rooks directly down the board
-        while(row + 1 != 8){
-            row++;
-            pos.add(new int[]{row, col});
-            int x = hasOpposingQueenRook(board, row, col, king.getColor());
-            if (x == 1) {
-                this.positions = toArrayMoves(pos);
-                return true;
-            } else if (x == -1) {
-                break;
-            }
-        }
-
-        pos = new ArrayList<>();
-        row = king.getRow();
-
-        // Checking for any Queens or Rooks directly up the board
-        while(row - 1 != -1) {
-            row--;
-            pos.add(new int[]{row, col});
-            int x = hasOpposingQueenRook(board, row, col, king.getColor());
-            if (x == 1) {
-                this.positions = toArrayMoves(pos);
-                return true;
-            } else if (x == -1) {
-                break;
-            }
-        }
-        return false;
-    }
-
-
-    /**
-     * This is a helper method used to determine whether the there is an opposing Queen or Rook
-     * chess piece located at the board in the given position.
-     *
-     * @return -  1, if there is an opposing Queen or Rook at the given position
-     *         -  0, if there is no chess piece at this given position
-     *         - -1, if there is a chess piece at this given position, but it is not an opposing one
-     */
-    private int hasOpposingQueenRook(ChessPiece[][] board, int row, int col, String color) {
-        if (board[row][col] != null) {
-            if (!(board[row][col].getColor().equals(color)) &&
-                    (board[row][col] instanceof Queen || board[row][col] instanceof Rook)) {
-                return 1;
-            } else {
-                return -1;
-            }
-        }
-        return 0;
-    }
-
-    /**
-     * This method is used to check is there is an opposing Queen or Bishop that can capture
-     * the king through a diagonal.
-     *
-     * @return - true, if there is either an opposing Queen or Bishop that can capture the King
-     *         - false, otherwise
-     */
-    private boolean checkDiagonals (King king, ChessPiece[][] board){
-        int col = king.getColumn();
-        int row = king.getRow();
-        ArrayList<int[]> pos = new ArrayList<>();
-
-        // Check for Queens or Bishops within up-right diagonal
-        while (col + 1 != 8 && row - 1 != -1) {
-            col++;
-            row--;
-            pos.add(new int[]{row, col});
-            int x = hasOpposingQueenBishop(board, row, col, king.getColor());
-            if (x == 1) {
-                this.positions = toArrayMoves(pos);
-                return true;
-            } else if (x == -1) {
-                break;
-            }
-        }
-
-        pos = new ArrayList<>();
-        col = king.getColumn();
-        row = king.getRow();
-
-        // Check for Queens or Bishops within up-Left Diagonal
-        while (col - 1 != -1 && row - 1 != -1) {
-            col--;
-            row--;
-            pos.add(new int[]{row, col});
-            int x = hasOpposingQueenBishop(board, row, col, king.getColor());
-            if (x == 1) {
-                this.positions = toArrayMoves(pos);
-                return true;
-            } else if (x == -1) {
-                break;
-            }
-        }
-
-        pos = new ArrayList<>();
-        col = king.getColumn();
-        row = king.getRow();
-
-        // Check for Queens or Bishops within down-Right Diagonal
-        while (col + 1 != 8 && row + 1 != 8) {
-            col++;
-            row++;
-            pos.add(new int[]{row, col});
-            int x = hasOpposingQueenBishop(board, row, col, king.getColor());
-            if (x == 1) {
-                this.positions = toArrayMoves(pos);
-                return true;
-            } else if (x == -1) {
-                break;
-            }
-        }
-
-        pos = new ArrayList<>();
-        col = king.getColumn();
-        row = king.getRow();
-
-        // Check for Queens or Bishops within down-Left Diagonal
-        while (col - 1 != -1 && row + 1 != 8) {
-            col--;
-            row++;
-            pos.add(new int[]{row, col});
-            int x = hasOpposingQueenBishop(board, row, col, king.getColor());
-            if (x == 1) {
-                this.positions = toArrayMoves(pos);
-                return true;
-            } else if (x == -1) {
-                break;
-            }
-        }
-        return false;
-    }
-
-    /**
-     * This is a helper method used to determine whether the there is an opposing Queen or Bishop
-     * chess piece located at the board in the given position.
-     *
-     * @return -  1, if there is an opposing Queen or Bishop at the given position
-     *         -  0, if there is no chess piece at this given position
-     *         - -1, if there is a chess piece at this given position, but it is not an opposing one
-     */
-    private int hasOpposingQueenBishop(ChessPiece[][] board, int row, int col, String color) {
-        if (board[row][col] != null) {
-            if (!(board[row][col].getColor().equals(color)) &&
-                    (board[row][col] instanceof Queen || board[row][col] instanceof Bishop)) {
-                return 1;
-            } else {
-                return -1;
-            }
-        }
-        return 0;
-    }
-
-    /**
-     * This method returns true if there are any opposing Knight pieces that
-     * are able to capture the King, and false otherwise.
-     *
-     * @return - true, if there opposing knight pieces that can capture the king
-     *         - false, otherwise
-     */
-    private boolean checkKnight(King king, ChessPiece[][] board){
-        int row = king.getRow();
-        int col = king.getColumn();
-
-        // Opposing Knight two spaces UP and one space LEFT
-        if (row - 2 > -1 && col - 1 > -1){
-            if(hasOpposingKnight(board, row - 2, col - 1, king.getColor())){
-                this.positions = new int[][]{{row - 2, col - 1}};
-                return true;
-            }
-        }
-
-        // Opposing Knight two spaces UP and one space RIGHT
-        if (row - 2 > -1 && col + 1 < 8 ){
-            if(hasOpposingKnight(board, row - 2, col + 1, king.getColor())){
-                this.positions = new int[][]{{row - 2, col + 1}};
-                return true;
-            }
-        }
-
-        // Opposing Knight two spaces DOWN and one space LEFT
-        if (row + 2 < 8 && col - 1 > -1){
-            if(hasOpposingKnight(board, row + 2, col - 1, king.getColor())){
-                this.positions = new int[][]{{row + 2, col - 1}};
-                return true;
-            }
-        }
-
-        // Opposing Knight two spaces DOWN and one space RIGHT
-        if (row + 2 < 8 && col + 1 < 8){
-            if(hasOpposingKnight(board, row + 2, col + 1, king.getColor())){
-                this.positions = new int[][]{{row + 2, col + 1}};
-                return true;
-            }
-        }
-
-        // Opposing Knight two spaces LEFT and one space UP
-        if (row - 1 > - 1 && col - 2 > -1){
-            if(hasOpposingKnight(board, row - 1, col - 2, king.getColor())){
-                this.positions = new int[][]{{row - 1, col - 2}};
-                return true;
-            }
-        }
-
-        // Opposing Knight two spaces LEFT and one space DOWN
-        if (row + 1 < 8 && col - 2 > -1){
-            if(hasOpposingKnight(board, row + 1, col - 2, king.getColor())){
-                this.positions = new int[][]{{row + 1, col - 2}};
-                return true;
-            }
-        }
-
-        // Opposing knight two spaces RIGHT and one space UP
-        if (row - 1 > -1 && col + 2 < 8){
-            if(hasOpposingKnight(board, row - 1, col + 2, king.getColor())){
-                this.positions = new int[][]{{row - 1, col + 2}};
-                return true;
-            }
-        }
-
-        // Opposing knight two spaces RIGHT and one space DOWN
-        if (row + 1 < 8 && col + 2 < 8){
-            if(hasOpposingKnight(board, row + 1, col + 2, king.getColor())){
-                this.positions = new int[][]{{row + 1, col + 2}};
-                return true;
-            }
-        }
-
-        return false;
-    }
-
-    /**
-     * This is a helper method used to determine whether the there is an opposing Knight
-     * chess piece located at the board in the given position.
-     *
-     * @return - true, if there is an opposing knight piece
-     *         - false, otherwise
-     */
-    private boolean hasOpposingKnight(ChessPiece[][] board, int row, int col, String color){
-        if (board[row][col] != null) {
-            return !(board[row][col].getColor().equals(color)) &&
-                    board[row][col] instanceof Knight;
-        }
-        return false;
-    }
-
-    /**
-     * This method is used to determine whether there are any opposing pawn pieces that are in
-     * position to capture the King.
-     *
-     * @return - true, if there is an opposing Pawn piece that can capture the King,
-     *         - false, otherwise
-     */
-    private boolean checkPawn(King king, ChessPiece[][] board){
-        String color = king.getColor();
-        int row = king.getRow();
-        int col = king.getColumn();
-
-        if(color.equals("Black")){
-            // If the pawn is to the bottom Left
-            if(row + 1 < 8 && col - 1 > - 1){
-                if(hasOpposingPawn(board, row + 1, col - 1, king.getColor())){
-                    this.positions = new int[][]{{row + 1, col - 1}};
-                    return true;
-                }
-            }
-            // If the pawn is to the bottom Right
-            if(row + 1 < 8 && col + 1 < 8){
-                if(hasOpposingPawn(board, row + 1, col + 1, king.getColor())){
-                    this.positions = new int[][]{{row + 1, col + 1}};
-                    return true;
-                }
-            }
-        }
-        else{
-            // If the pawn is to the upper Left
-            if(row - 1 > -1 && col - 1 > - 1){
-                if(hasOpposingPawn(board, row - 1, col - 1, king.getColor())){
-                    this.positions = new int[][]{{row - 1, col - 1}};
-                    return true;
-                }
-            }
-            // If the pawn is to the upper Right
-            if(row - 1 > -1 && col + 1 < 8){
-                if(hasOpposingPawn(board, row - 1, col + 1, king.getColor())){
-                    this.positions = new int[][]{{row - 1, col + 1}};
-                    return true;
-                }
-            }
-        }
-        return false;
-    }
-
-    /**
-     * This is a helper method used to determine whether the there is an opposing Pawn
-     * chess piece located at the board in the given position.
-     *
-     * @return - true, if there is an opposing pawn in the given position
-     *         - false, otherwise
-     */
-    private boolean hasOpposingPawn(ChessPiece[][] board, int row, int col, String color){
-        if (board[row][col] != null) {
-            return !(board[row][col].getColor().equals(color)) &&
-                    board[row][col] instanceof Pawn;
-        }
-        return false;
     }
 }
